Ignore unmerged fields
